pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "datadog-healthcheck-deployer"
version = "0.1.0"
description = "A tool for deploying and managing DataDog health checks using configuration as code"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "fleXRPL", email = "info@flexrpl.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "datadog>=0.44.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "requests>=2.26.0",
    "jsonschema>=4.0.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "pylint>=2.17.0",
    "bandit>=1.7.0",
    "pre-commit>=2.17.0",
]

[project.urls]
Homepage = "https://github.com/fleXRPL/datadog-healthcheck-deployer"
Documentation = "https://github.com/fleXRPL/datadog-healthcheck-deployer/wiki"
Repository = "https://github.com/fleXRPL/datadog-healthcheck-deployer.git"
Issues = "https://github.com/fleXRPL/datadog-healthcheck-deployer/issues"

[project.scripts]
dd-healthcheck = "datadog_healthcheck_deployer.cli:main"

[tool.black]
line-length = 100
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=datadog_healthcheck_deployer.config,datadog_healthcheck_deployer.checks.http,datadog_healthcheck_deployer.core --cov-report=xml --cov-report=term-missing --cov-fail-under=90"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
branch = true
source = ["src"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]
fail_under = 80  # Temporarily lower this from 90 until we have more tests
ignore_errors = true 